+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [#5554](https://github.com/influxdata/influxdb/issues/5554): Can't run in alpine linux
 - [#6094](https://github.com/influxdata/influxdb/issues/6094): Ensure CREATE RETENTION POLICY and CREATE CONTINUOUS QUERY are idempotent in the correct way.
 - [#6092](https://github.com/influxdata/influxdb/issues/6092): Upgrading directly from 0.9.6.1 to 0.11.0 fails
+- [#6061](https://github.com/influxdata/influxdb/issues/6061): [0.12 / master] POST to /write does not write points if request has header 'Content-Type: application/x-www-form-urlencoded'
 
 ## v0.11.0 [2016-03-22]